Refactor common block setup tasks
